+++ packages/entities/entities-certificates/src/components/CACertificateList.vue
@@ -101,6 +101,13 @@
               />
             </div>
           </template>
+
+          <template
+            v-if="config?.isControlPlaneGroup"
+            #message
+          >
+            {{ t('ca-certificates.list.empty_state_v2.group') }}
+          </template>
         </EntityEmptyState>
       </template>