+++ packages/entities/entities-routes/src/components/RouteList.vue
@@ -105,6 +105,13 @@
               />
             </div>
           </template>
+
+          <template
+            v-if="config?.isControlPlaneGroup"
+            #message
+          >
+            {{ t('routes.list.empty_state_v2.group') }}
+          </template>
         </EntityEmptyState>
       </template>