+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.10.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.10.10...@kong-ui-public/entities-routes@3.10.11) (2025-03-07)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.10.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.10.9...@kong-ui-public/entities-routes@3.10.10) (2025-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
 ## [3.10.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.10.8...@kong-ui-public/entities-routes@3.10.9) (2025-03-04)
 
 **Note:** Version bump only for package @kong-ui-public/entities-routes